Allow centos 8-stream to be mirrored
Since the current CentOS 8 repositories are now under http://mirror.centos.org/centos/8-stream/ we need to be able to download/mirror this repository.
When using `8-stream` the script fails as it tries to evaluate stream as an integer making a range from `8` to `stream` which doesn't make sense obviously.

I did this locally and was able to mirror the `8-stream` repos. Just figured someone might also need that :smile: 
Let me know if and where to add tests.